New design

+++ APPSenSoutenance/Views/Parametre/frmAnneeAcademique.cs
@@ -1,6 +1,6 @@
-﻿using APPSenSoutenance.Models;
+using APPSenSoutenance.Models;
+using APPSenSoutenance.Shared;
 using System;
-using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
 
@@ -11,24 +11,16 @@ namespace APPSenSoutenance.Views.Parametre
         public frmAnneeAcademique()
         {
             InitializeComponent();
+            this.Load += new EventHandler(frmAnneeAcademique_Load);
         }
 
         BdSenSoutenanceContext db = new BdSenSoutenanceContext();
 
-        private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
-        {
-            dgAnneeAcademique.DataSource = db.AnneeAcademiques.ToList();
-        }
-
         private void frmAnneeAcademique_Load(object sender, EventArgs e)
         {
-            // Appliquer le style des lignes alternées au DataGridView
-            dgAnneeAcademique.AlternatingRowsDefaultCellStyle.BackColor = Color.FromArgb(236, 240, 245);
-            dgAnneeAcademique.AlternatingRowsDefaultCellStyle.ForeColor = Color.FromArgb(52, 73, 94);
-            dgAnneeAcademique.DefaultCellStyle.BackColor = Color.White;
-            dgAnneeAcademique.DefaultCellStyle.ForeColor = Color.FromArgb(52, 73, 94);
-
-            // Charger les données
+            this.BackColor = DarkTheme.BgPrincipal;
+            this.ForeColor = DarkTheme.TextPrimary;
+            DarkTheme.StyleDataGridView(dgAnneeAcademique);
             Effacer();
         }
 
@@ -39,8 +31,10 @@ namespace APPSenSoutenance.Views.Parametre
             dgAnneeAcademique.DataSource = db.AnneeAcademiques.ToList();
             txtLibelleAnneeAcademique.Focus();
         }
+
         private void btnAdd_Click(object sender, EventArgs e)
         {
+            if (string.IsNullOrWhiteSpace(txtLibelleAnneeAcademique.Text)) { MessageBox.Show("Le libellé est requis.", "Attention"); return; }
             AnneeAcademique anneeAcademique = new AnneeAcademique
             {
                 LibelleAnneeAcademique = txtLibelleAnneeAcademique.Text,
@@ -53,6 +47,7 @@ namespace APPSenSoutenance.Views.Parametre
 
         private void btnEdit_Click(object sender, EventArgs e)
         {
+            if (dgAnneeAcademique.CurrentRow == null) return;
             int? id = int.Parse(dgAnneeAcademique.CurrentRow.Cells[0].Value.ToString());
             AnneeAcademique anneeAcademique = db.AnneeAcademiques.Find(id);
             anneeAcademique.LibelleAnneeAcademique = txtLibelleAnneeAcademique.Text;
@@ -63,6 +58,7 @@ namespace APPSenSoutenance.Views.Parametre
 
         private void btnRemove_Click(object sender, EventArgs e)
         {
+            if (dgAnneeAcademique.CurrentRow == null) return;
             int? id = int.Parse(dgAnneeAcademique.CurrentRow.Cells[0].Value.ToString());
             AnneeAcademique anneeAcademique = db.AnneeAcademiques.Find(id);
             db.AnneeAcademiques.Remove(anneeAcademique);
@@ -77,7 +73,6 @@ namespace APPSenSoutenance.Views.Parametre
                 MessageBox.Show("Veuillez sélectionner une ligne.", "Attention");
                 return;
             }
-
             txtLibelleAnneeAcademique.Text = dgAnneeAcademique.CurrentRow.Cells[1].Value.ToString();
             txtAnneeAcademiqueVal.Text = dgAnneeAcademique.CurrentRow.Cells[2].Value.ToString();
         }
