+++ APPSenSoutenance/Views/Parametre/frmSession.cs
@@ -1,4 +1,4 @@
-﻿using AppSenSoutenance.Shared;
+using APPSenSoutenance.Shared;
 using APPSenSoutenance.Models;
 using System;
 using System.Data;
@@ -6,29 +6,32 @@ using System.Linq;
 using System.Windows.Forms;
 using Session = APPSenSoutenance.Models.Session;
 
-namespace AppSenSoutenance.View.Paramètre
+namespace APPSenSoutenance.Views.Parametre
 {
     public partial class frmSession : Form
     {
         public frmSession()
         {
             InitializeComponent();
+            this.Load += new EventHandler(frmSession_Load);
         }
 
         BdSenSoutenanceContext db = new BdSenSoutenanceContext();
-
         FillerList filler = new FillerList();
+
         private void frmSession_Load(object sender, EventArgs e)
         {
+            this.BackColor = DarkTheme.BgPrincipal;
+            DarkTheme.StyleDataGridView(dgSession);
             Effacer();
         }
 
-        private void btnAdd_Click_1(object sender, EventArgs e)
+        private void btnAdd_Click(object sender, EventArgs e)
         {
+            if (cbbAnneeAcademique.SelectedValue == null) { MessageBox.Show("Sélectionnez une année académique.", "Attention"); return; }
             Session session = new Session();
             session.LibelleSession = txtSession.Text;
             session.IdAnneeAcademique = int.Parse(cbbAnneeAcademique.SelectedValue.ToString());
-
             db.Sessions.Add(session);
             db.SaveChanges();
             Effacer();
@@ -37,7 +40,6 @@ namespace AppSenSoutenance.View.Paramètre
         private void Effacer()
         {
             txtSession.Clear();
-            cbbAnneeAcademique.SelectedValue = "";
             dgSession.DataSource = db.Sessions.ToList();
             cbbAnneeAcademique.DataSource = filler.FillAnneeAcademique();
             cbbAnneeAcademique.DisplayMember = "Text";
@@ -47,6 +49,7 @@ namespace AppSenSoutenance.View.Paramètre
 
         private void btnSelect_Click(object sender, EventArgs e)
         {
+            if (dgSession.CurrentRow == null) return;
             int? id = int.Parse(dgSession.CurrentRow.Cells[0].Value.ToString());
             Session session = db.Sessions.Find(id);
             txtSession.Text = session.LibelleSession;
@@ -55,6 +58,7 @@ namespace AppSenSoutenance.View.Paramètre
 
         private void btnEdit_Click(object sender, EventArgs e)
         {
+            if (dgSession.CurrentRow == null) return;
             int? id = int.Parse(dgSession.CurrentRow.Cells[0].Value.ToString());
             Session session = db.Sessions.Find(id);
             session.LibelleSession = txtSession.Text;
@@ -65,6 +69,7 @@ namespace AppSenSoutenance.View.Paramètre
 
         private void btnRemove_Click(object sender, EventArgs e)
         {
+            if (dgSession.CurrentRow == null) return;
             int? id = int.Parse(dgSession.CurrentRow.Cells[0].Value.ToString());
             Session session = db.Sessions.Find(id);
             db.Sessions.Remove(session);
@@ -76,20 +81,12 @@ namespace AppSenSoutenance.View.Paramètre
         {
             var liste = db.Sessions.ToList();
             if (!string.IsNullOrEmpty(txtRSession.Text))
-            {
                 liste = liste.Where(s => s.LibelleSession.Contains(txtRSession.Text)).ToList();
-            }
-
-            if (txtRanneeAcademique.Text != "")
-            {
-                liste = liste.Where(s => s.AnneeAcademique.LibelleAnneeAcademique.Contains(txtRanneeAcademique.Text)).ToList();
-            }
+            if (!string.IsNullOrEmpty(txtRanneeAcademique.Text))
+                liste = liste.Where(s => s.AnneeAcademique != null && s.AnneeAcademique.LibelleAnneeAcademique.Contains(txtRanneeAcademique.Text)).ToList();
             dgSession.DataSource = liste;
         }
 
-        private void txtSession_TextChanged(object sender, EventArgs e)
-        {
-
-        }
+        private void txtSession_TextChanged(object sender, EventArgs e) { }
     }
 }
